cleanup up code
